Move CBanEntry from net.h to banentry.h|.cpp
Includes addition of doxygen comments and enforcing clang-format.
